requirements.txt
clip==1.0
gradio==4.21.0
numpy==1.21.5
paddleocr==2.7.2
paddleocr.egg==info
Pillow==9.2.0
pymongo==4.6.2
PyYAML==6.0.1
tqdm==4.64.1
ftfy
regex
torch
torchvision
git+https://github.com/openai/CLIP.git